+++ bins/enigma-daemon/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod config;
+pub const DAEMON_VERSION: &str = env!("CARGO_PKG_VERSION");